wip refactor old code
